*moved cumulus runtime config to xcm_config module
